### 2019-08-02 #1625 @whedon
[PRE REVIEW]: SMBLtoODEpy: A software program for converting SBML models into ODE models in Python
**Submitting author:** @ashleefv (<a href="http://orcid.org/0000-0001-9059-5703">Ashlee N. Ford Versypt</a>)
**Repository:** <a href="https://github.com/SMRuggiero/sbmltoodepy" target ="_blank">https://github.com/SMRuggiero/sbmltoodepy</a>
**Version:** v.1.0.2
