+++ paper.bib
@@ -11,7 +11,7 @@
 }
 @article{chelliah2015,
     author = {Chelliah, V. and Juty, N. and Ajmera, I. and Ali, R. and Dumousseau, M. and Glont, M. and Hucka, M. and Jalowicki, G. and Keating, S. and Knight-Schrijver, V. and Lloret-Villas, A. and Natarajan, K. N. and Pettit, J.-B. and Rodriguez, N. and Schubert, M. and Wimalaratne, S. M. and Zhao, Y. and Hermjakob, H. and Le Novère, N. and Laibe, C.},
-    title = "{BioModels: Ten-year anniversary}},
+    title = {{BioModels: Ten-year anniversary}},
     journal = {Nucleic Acids Research},
     volume = {43},
     number = {D1},
