### 2019-08-12 #1643 @whedon
[REVIEW]: SMBLtoODEpy: A software program for converting SBML models into ODE models in Python
**Submitting author:** @ashleefv (<a href="http://orcid.org/0000-0001-9059-5703">Ashlee N. Ford Versypt</a>)
**Repository:** <a href="https://github.com/SMRuggiero/sbmltoodepy" target ="_blank">https://github.com/SMRuggiero/sbmltoodepy</a>
**Version:** v1.0.3
