### 2019-08-12 #1625 @mgymrek
Please head over to #1643 to perform the review. If you are new to reviewing for JOSS, please let me know if you have any questions.
