### 2019-08-13 SMRuggiero/sbmltoodepy#1 @SirSharpest
JOSS submission comments
Regarding JOSS review: [https://github.com/openjournals/joss-reviews/issues/1643](https://github.com/openjournals/joss-reviews/issues/1643)

To keep the JOSS review page as clean as possible I'll be making my comments here.

I've just begun reading though your software package, many of the comments I'll make here are suggestions - anything which I feel is a serious concern I will state as such.

I'll also make it clear when I've finished.

Finally, many thanks for making a contribution to open source software! :)
