### 2019-08-13 SMRuggiero/sbmltoodepy#1 @AoifeHughes
# General comments
- It would be useful to state which range of python 2X / 3.X this software is supporting.
- A brief line on how to install from source would be useful - I see that one is given for installation from pip

# Testing
- Effort has been made to provide this software with testing [TestingPackage Class](https://github.com/SMRuggiero/sbmltoodepy/blob/8abcff615b4ce31e77951ac08eb5fcdddb5af8e3/sbmltoodepy/utilities.py#L77). However it might be more useful if a full unit-testing frame work was used, for example [PyTest](https://docs.pytest.org/en/latest/) (though other frameworks are available). I would say this is quite important, especially for isolating any potential bugs which could crop up across different python versions or software versions.

## Software checked on:

I wouldn't expect python 2.7 to be support, although some holdouts still use it in the scientific community. It would be useful to explicity state which python versions are "officially" supported, as mentioned previously

- Python 3.6 (works)
- Python 3.7 (works)
- Python 2.7 (won't install)

# Software style
- A very minor detail - some practises are not ideal in the formatting of the software and could lead to large errors. For example the use of `from x import *`can lead to unexpected behaviour. A style guide / linter would be helpful [pep8](https://www.python.org/dev/peps/pep-0008/)

# Running the provided Tutorial.md
- The line `pip install sbmltoodepy` can give 2 pieces of confusion.
  1. It will install a potentially different version than the repo cloned in the previous lines
  2. It needs a '!' to be run from jupyter e.g. `!pip install sbmltoodepy`
- Both examples given run and give reproducible results
- It might be worth hosting an .ipynb file in the git repo so that others can directly run when they clone it.
