### 2019-08-13 SMRuggiero/sbmltoodepy#1 @AoifeHughes
# Community guidelines
- If you see the scope for others' to make pull requests or suggest features then it may be an idea to provide some instructions as to how best do this.

# Paper clarifications
- The authors have made comparisons to existing software soultions, although I am unsable to see where or how these are evaluated to their proposed method? Perhaps I have misunderstood, but I think it would be useful to focus on this evaluation and show how this software package improves upon existing methods in more detail in the submitted paper.
