### 2019-08-13 #1643 @SirSharpest
@ashleefv I have made my comments as a new issue on the relevant github page: [https://github.com/SMRuggiero/sbmltoodepy/issues/1](https://github.com/SMRuggiero/sbmltoodepy/issues/1)

Most are minor, and possibly non-issues. Though, I feel that the clarifications I mentioned for the manuscript would be important to readers. As would being explicit in which python versions are supported. Finally, I feel strongly about unit-testing being done to software engineering standards, but I understand that this has at least been addressed using their own method.

@mgymrek if there is anything else, prior to author response, please let me know and I'll promptly proceed with.
