### 2019-08-21 SMRuggiero/sbmltoodepy#2 @marouenbg
JOSS reviewer 2 comment 1
Review openjournals/joss-reviews#1643
sbmltoode is a great tool for fast conversion and simulation of ode models, the conversion to json strings is a great feature to have.

The software installs correctly on python 3.7, tests correctly with COPASI results, tutorials are clear. If possible, it would be nice to clarify how sbmltoode preserves model components potentially through applying standards like [sed-ml](https://sed-ml.org/) and [kisao](https://co.mbine.org/standards/kisao). From a user perspective, the main goal of a converter would be to maintain the original model specification.
Also, if possible, a sentence in the text to compare COPASI and sbmltoode.
