### 2019-08-21 SMRuggiero/sbmltoodepy#3 @marouenbg
JOSS reviewer 2 comment 2
Review openjournals/joss-reviews#1643

The package integrates a test suite to compare with COPASI results as a ground truth. If possible, you can consider continuous integration through a CI server like travis-ci. Recently, Github enabled CI, have a look [here](https://help.github.com/en/articles/setting-up-continuous-integration-using-github-actions).
CI would be a great addition to your package especially that you have the tests written and the model files in the repo. Alternatively, you can consider a [code ocean capsule](https://codeocean.com/) that includes entire OS with hardware for extended cross-compatiblity testing.
