### 2019-08-21 SMRuggiero/sbmltoodepy#4 @marouenbg
JOSS reviewer 2 comment 3
Review openjournals/joss-reviews#1643

my last minor comment has to do with the documentation website which already covers automatically the functions description. If possible, I suggest to organize it as introduction, full installation guide, troubleshooting, function description (expand if possible beyond the argument specification), tutorials (paste your tutorials here), and changelog of the documentation.
