### 2019-08-26 SMRuggiero/sbmltoodepy#1 @ashleefv
## Running the provided Tutorial.md

We addressed the comments by
- clarifying the case when cloning the repo and pip install make sense together (use on google colab)
- fixing pip install command
- hosting a copy of SBMLtoODEpy.ipynb in the repo
