+++ sbmltoodepy/parse.py
@@ -4,10 +4,10 @@ Created on Tue Jul  3 12:12:05 2018
 
 @author: Steve
 """
-from libsbml import *
+import libsbml
 import re
 import sys
-from sbmltoodepy.dataclasses import *
+import sbmltoodepy.dataclasses
 
 def ParseParameterAssignment(parameterIndex, parameter):
     #if parameter.isSetValue() and parameter.isSetId():
@@ -18,7 +18,7 @@ def ParseParameterAssignment(parameterIndex, parameter):
 #        #print("Parameter," + str(parameterIndex + 1) + "\n" + str(parameterId) + "\nValue\n" + str(parameterValue))
 #        outputFile.write(str(parameterName) + ";" + str(parameterValue) + "\n")
 #        return
-    newParameter = ParameterData()
+    newParameter = sbmltoodepy.dataclasses.ParameterData()
     if parameter.isSetName():
         newParameter.name = parameter.getName()
     else:
@@ -46,10 +46,10 @@ def ParseRule(ruleIndex, rule):
 
 #        outputFile.write("Rule; " + ruleId + "; Algebraic; " + ruleName + "\n")
     if rule.isAssignment():
-        newRule = AssignmentRuleData()
+        newRule = sbmltoodepy.dataclasses.AssignmentRuleData()
 #        outputFile.write("Rule; " + ruleId + "; Assignment; " + ruleName + "\n")
     if rule.isRate():
-        newRule = RateRuleData()
+        newRule = sbmltoodepy.dataclasses.RateRuleData()
 #        outputFile.write("Rule; " + ruleId + "; Rate; " + ruleName + "\n")
     
     if rule.isSetName():
@@ -63,7 +63,7 @@ def ParseRule(ruleIndex, rule):
         newRule.Id = str(ruleIndex + 1)
     
     newRule.variable = rule.getVariable()
-    newRule.math = formulaToL3String(rule.getMath())
+    newRule.math = libsbml.formulaToL3String(rule.getMath())
 
 
         
@@ -77,7 +77,7 @@ def ParseRule(ruleIndex, rule):
         
 def ParseSpecies(speciesIndex, species):
 
-    newSpecies = SpeciesData()
+    newSpecies = sbmltoodepy.dataclasses.SpeciesData()
     
     if species.isSetName():
         newSpecies.name = species.getName()
@@ -110,7 +110,7 @@ def ParseSpecies(speciesIndex, species):
     
         
 def ParseReaction(reactionIndex, reaction):
-    newReaction = ReactionData()
+    newReaction = sbmltoodepy.dataclasses.ReactionData()
     if reaction.isSetIdAttribute():
         newReaction.Id = reaction.getIdAttribute()
     else:
@@ -148,7 +148,7 @@ def ParseReaction(reactionIndex, reaction):
     rateLawObject = reaction.getKineticLaw()
     
     if rateLawObject.getMath() != None:
-        newReaction.rateLaw = formulaToL3String(rateLawObject.getMath())
+        newReaction.rateLaw = libsbml.formulaToL3String(rateLawObject.getMath())
 #        outputFile.write(formulaToString(rateLaw.getMath()) + "\n")
     else:
         raise Exception("Rate law defined by plugin that is not currently supported")
@@ -171,7 +171,7 @@ def ParseReaction(reactionIndex, reaction):
         
 def ParseCompartment(compartmentIndex, compartment):
     
-    newCompartment = CompartmentData()
+    newCompartment = sbmltoodepy.dataclasses.CompartmentData()
     
     newCompartment.Id = compartment.getId()
     
@@ -201,7 +201,7 @@ def ParseCompartment(compartmentIndex, compartment):
     
 def ParseFunction(functionIndex, function):
     
-    newFunction = FunctionData()
[…]
