### 2019-08-29 d596af9
Added bit about installing from source
