### 2019-09-01 #1643 @ashleefv
Alright. We've been posting responses on the github repository issues. We'll also consolidate them and post here when they are completed (should be later this week).

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu<mailto:ashleefv@okstate.edu>
Ford Versypt Lab Website<https://sites.google.com/site/ashleefordversypt/>
