### 2019-09-02 e8d79e0
Edited paper to go into more detail about comparison to SBFC and COPASI.
