+++ paper.bib
@@ -176,4 +176,11 @@ doi={10.1093/bioinformatics/btl485},
 url={https://doi.org/10.1093/bioinformatics/btl485}
 }
 
-
+@article{hindmarsh1983,
+  title={ODEPACK, a systematized collection of ODE solvers},
+  author={Hindmarsh, Alan C},
+  journal={Scientific computing},
+  pages={55--64},
+  year={1983},
+  publisher={North-Holland}
+}
