### 2019-09-02 SMRuggiero/sbmltoodepy#1 @AnabelSMRuggiero
# General Comments

We now specify that the supported versions of python are 3.5, 3.6, and 3.7.

Added comment to README.md about installing from source.

# Software style

Removed all of the "import *" statements. While the updated code on this github repo has been updated, we plan on updating the package on PyPI once all reviewer comments have been addressed.

# Paper Clarifications

In our paper, we compare our work to two other pieces of software.

COPASI- The comparison was intended simply to validate the results of the models generated by SBMLtoODEpy. COPASI uses ODEPACK for its Time Course calculations, which is the same library that scipy's odeint function uses. If the calculations agree, then that suggests SBMLtoODEpy has generated the correct equations.  Furthermore, the targeted niche for SBMLtoODEpy is not the same as COPASI.

Systems Biology Format Converter (SBFC)- SBFC is a better software to compare SBMLtoODEpy to. SBFC converts SBML models into various programing and modeling languages, but notably excludes Python. This gap SBFC's functionality is specifically the target niche of SBMLtoODEpy. Additionally, SBMLtoODEpy is designed to generate code that can be incorporated into other python projects. The Octave/MATLAB scripts generated by SBFC implement the model that neither accept arguments nor return any values. SBMLtoODEpy generates code that implements the model as a class with most methods and members intended to be accessible to other code.

The paper has been updated to better convey the scope and intent of the comparisons.
