### 2019-09-02 c3fa1dc
Updated documentation and incorportated readme and tutortial markdown files into documentation.
