+++ build/lib/sbmltoodepy/modulegeneration.py
@@ -1,0 +1,546 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul  3 15:50:45 2018
+
+@author: Steve
+"""
+
+import re
+import numpy as np
+import sys
+
+
+            
+def GenerateModel(modelData, outputFilePath, objectName = 'SBMLmodel'):
+    """
+    This function takes model data, either from ParseSBMLFIle() or imported from a .json file,
+    and generates a Python file containing a class that implements the SBML model.
+    
+    Parameters
+    ----------
+    modelData : ModelData
+        An object containing all of the model components and values.
+		
+    outputFilePath : str
+        The desired file path of the resulting python file.
+		
+    objectName : str
+        The name of the class defined in the resulting python file.
+		
+    Returns
+    -------
+    None
+    """
+    #The library mathFuncs serves to both only allow functions supported
+    #functions in SBML/user defined functions, but also the python equivalent
+    
+    np.set_printoptions(threshold=sys.maxsize)
+    
+    
+    outputFile = open(outputFilePath, "w")
+
+    parameters = modelData.parameters
+    compartments = modelData.compartments
+    species = modelData.species
+    reactions = modelData.reactions
+    functions = modelData.functions
+    
+    assignmentRules = modelData.assignmentRules
+    rateRules = modelData.rateRules
+    initialAssignments = modelData.initialAssignments
+    
+    mathFuncs = {'abs' : 'abs',
+                 'max' : 'max',
+                 'min' : 'min',
+                 'pow' : 'pow',
+                 'exp' : 'math.exp',
+                 'floor' : 'np.floor',
+                 'ceiling' : 'math.ceil',
+                 'exp' : 'math.exp',
+                 'ln' : 'math.log',
+                 'log' : 'math.log10',
+                 'factorial' : 'math.factorial',
+                 'sqrt' : 'math.sqrt',
+                 
+                 'eq' : 'operator.eq',
+                 'neq' : 'operator.ne',
+                 'gt' : 'operator.gt',
+                 'lt' : 'operator.lt',
+                 'geq' : 'operator.ge',
+                 'leq' : 'operator.le',
+                 
+                 'and' : 'operator.and_',
+                 'or' : 'operator.or_',
+                 'xor' : 'operator.xor_',
+                 'not' : 'operator.not_',
+                 
+                 'sin' : 'np.sin',
+                 'cos' : 'np.cos',
+                 'tan' : 'np.tan',
+                 'sec' : '1/np.cos',
+                 'csc' : '1/np.sin',
+                 'cot' : '1/np.tan',
+                 'sinh' : 'np.sinh',
+                 'cosh' : 'np.cosh',
+                 'tanh' : 'np.tanh',
+                 'sech' : '1/np.cosh',
+                 'csch' : '1/np.sinh',
+                 'coth' : '1/np.tanh',
+                 'arcsin' : 'np.arcsin',
+                 'arccos' : 'np.arccos',
+                 'arctan' : 'np.arctan',
+                 'arcsinh' : 'np.arcsinh',
+                 'arccosh' : 'np.arccosh',
+                 'arctanh' : 'np.arctanh',
+                 
+                 'true' : 'True',
+                 'false' : 'False',
+                 'notanumber' : 'np.nan',
+                 'pi' : 'np.pi',
+                 'infinity' : 'np.inf',
+                 'exponentiale' : 'np.e',
+                 'piecewise' : 'sbmltoodepy.modelclasses.Piecewise'
+                 } 
+    #Add in user defined functions
+#    for function in functions:
+#        mathFuncs[function] = "self." + function
+		
+    #Set up stoichCoeffMat, a matrix of stoichiometric coefficients for solving the reactions
+    reactantCounter = 0
+    reactantIndex = {}
+    reactionCounter = 0
[…]
