+++ build/lib/sbmltoodepy/modelclasses.py
@@ -1,0 +1,389 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Dec 27 17:28:34 2018
+
+@author: Steve
+"""
+
+import warnings
+
+class Compartment:
+
+#    members:
+#        size (float) - This value represents the size of the shape. 
+#            Whether this is a volume or an area is determined by the compartments dimensionality.
+#            If the compartment is set to be constant, assigning a value to this compartment fails
+#            and a warning is raised.
+#		
+#        dimensionality (float or int) - The number of spatial dimensions this compartment exists in.
+#            Not necessarily a whole number.
+#            
+#        metadata (SBMLMetadata) - Information about the model component stored in the SBML file
+#        
+#        _constant (boolean) - 
+#            
+#        
+#	
+#    methods:
+#        __init__(self, size, dimensionality[, isConstant = False, metadata = None])
+#            arguements:
+#                size (float) - Sets the value of the Size member
+#				
+#                dimensionality (float or int) - Sets the value of the dimensionality member
+#                
+#                isConstant (boolean)
+#				
+#            returns:
+#                An instance of this class
+    """This class represents a compartment which is one of the model elements in SBML.
+    
+    Each species is associated with one compartment by referencing the appropriate 
+    compartment with its own compartment member.
+    
+    Parameters
+    ----------
+    size : float
+        This value represents the size of the shape. 
+        Whether this is a volume or an area is determined by the compartments dimensionality.
+        If the compartment is set to be constant, assigning a value to this compartment fails
+        and a warning is raised.
+	
+    dimensionality : int or float
+        The number of spatial dimensions this compartment exists in. This is not limited to whole numbers.
+            
+    isConstant : bool
+        If true, the compartment will be set to constant and attempts to change the size attribute will raise warnings.
+            
+    metadata : SBMLMetadata
+        An object containing information about the model component stored in the SBML file
+        
+    Attributes
+    ----------
+    dimensionality : float
+        The dimensionality of the compartment
+    metadata : SBMLMetadata
+        Additional information about the compartment contained in the SBML model
+    """
+    
+    def __init__(self, size, dimensionality, isConstant = False, metadata = None):
+
+        self._size = size
+        self.dimensionality = dimensionality
+        self._constant = isConstant
+        if metadata:
+            self.metadata = metadata
+            self.metadata._parentObject = self
+        else:
+            self.metadata = None
+            
+    @property
+    def size(self):
+        """float: The size of the comparment. This is used by species to calculate their concentrations."""
+        return self._size
+        
+    @size.setter
+    def size(self, size):
+        if not self._constant:
+            self._size = size
+        else:
+           warnings.warn("Attempted to set a constant parameter", Warning) 
+        
+
+#Currently unused
+class Reaction:
+    
+    def __init__(self, metadata = None):
+        
+        if metadata:
+            self.metadata = metadata
+            self.metadata._parentObject = self
+        else:
+            self.metadata = None
+        return
+        
+    
+        
+        
+class Species:
+    """This class represents a species as defined by SBML.
+    
+    Parameters
+    ----------
+    value : float
+        Sets either the value of conentration or amount, depending on
+        the concentrationIn value.
+
+    valueType : str
+        If valueType is equal to"Concentration", the amount input is assigned to the concentration member.
+        Otherwise, amount is assigned to the amount member.
+				
+    compartment : Compartment
+        Sets the compartment member to reference this input.
+    
+    hasOnlySubstanceUnits : bool
+        Determines if the species concentration or amount is used is rate laws
+				
+    constant : bool
[…]
