+++ source/conf.py
@@ -33,7 +33,8 @@ sys.path.insert(0, os.path.abspath('../sbmltoodepy'))
 extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.napoleon',
-    'sphinx.ext.mathjax'
+    'sphinx.ext.mathjax',
+    'recommonmark'
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -42,8 +43,8 @@ templates_path = ['_templates']
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
-# source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ['.rst', '.md']
+#source_suffix = '.rst'
 
 # The encoding of source files.
 #
