+++ build/lib/sbmltoodepy/__init__.py
@@ -1,0 +1,2 @@
+name = "sbmltopyode"
+from .utilities import ParseAndCreateModel
