+++ build/html/_sources/Parse.rst.txt
@@ -1,0 +1,18 @@
+====================
+ Parse and Generate
+====================
+
+.. autofunction:: utilities.TestPackage
+
+.. autofunction:: utilities.ParseAndCreateModel
+
+.. autofunction:: parse.ParseSBMLFile
+
+.. autofunction:: modulegeneration.GenerateModel
+
+.. toctree::
+	:maxdepth: 2
+	
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`

[…]
