### 2019-09-02 c401493
Reordered documentation sections and generated documentation based on files in top directory instead of a copy.
