+++ source/Tutorial.md
@@ -1,177 +1,0 @@
-# Tutorial
-
-## Interactive Jupyter Notebook working through the installation and a couple of example
-SBMLtoODEpy.ipynb
-
-Online version that doesn't require installation of Python:
-https://colab.research.google.com/drive/1yBscTNFwgliOUD10YhdWh-3HvXdj1YUF
-
-## Creating a Python model
-
-The fastest way to create a Python model using SBMLtoODEpy is to open a Python interpreter in the same folder as the SBML file. Then, run the following lines of code using the name of your model.
-
-```python
-import sbmltoodepy
-sbmltoodepy.ParseAndCreateModel("YourModelNameHere.xml")
-```
-
-If no errors occur, then a file with the same name as the SBML file but with the extension .py will be created in the same folder. To have a file created with a different name, you can use the outputFilePath keyword argument.
-
-```python
-sbmltoodepy.ParseAndCreateModel("YourModelNameHere.xml", outputFilePath = "PythonFile.py")
-```
-
-The new Python module defines a class named SBMLmodel. To customize the name of the class, use the className keyword argument.
-
-```python
-sbmltoodepy.ParseAndCreateModel("YourModelNameHere.xml", outputFilePath = "PythonFile.py", className = "ModelName")
-```
-
-The last optional keyword argument of ParseAndCreateModel() is jsonFilePath. If specified, a JSON file containing all of the model elements parsed from the SBML model will be created.
-
-```python
-sbmltoodepy.ParseAndCreateModel("YourModelNameHere.xml", jsonFilePath = "YourModelNameHere.json", outputFilePath = "PythonFile.py", className = "ModelName")
-```
-
-In reality, creating a Python model with SBMLtoODEpy is a two step process that can be broken up.
-
-```python
-dataOfModel = sbmltoodepy.parse.ParseSBMLFile("YourModelNameHere.xml")
-sbmltoodepy.modulegeneration.GenerateModel(dataOfModel, "PythonFile.py", objectName = "ModelName")
-```
-
-The function ParseAndCreateModel serves as a wrapper for two other functions, ParseSBMLFile and GenerateModel.
-ParseSBMLFile returns an instance of the ModelData class with all of the components pulled from the model.
-While you can use that instance as input to GenerateModel, there is another feature of SBMLtoODEpy that the ModelData class provides.
-The ModelData class has a method, DumpToJSON, that generates a JSON file with the model components.
-
-```python
-sbmltoodepy.dataclasses.ModelData.DumpToJSON(dataOfModel,"YourModelNameHere.json")
-```
-
-A JSON file created from the DumpToJSON method can be used a create a new instance of the ModelData class.
-
-```python
-newInstance = sbmltoodepy.dataclasses.ModelData.LoadFromJSON("YourModelNameHere.json")
-```
-
-A possible use case for this would be to generate a JSON file from an SBML model, change the JSON file, which is more human readable than an SBML file, and use the changed JSON file to create the Python model.
-
-
-## Exploring a newly created Python model
-
-After generating the Python implementation of a model, you can now import the new class and instantiate it.
-
-```python
-from PythonFile import ModelName
-modelInstance = ModelName()
-```
-
-The main components that make up the state of an SBML are compartments, parameters, and species. Each of these are stored in a dictionary that is a member of the model class. These members are named c, p,and s repectively.
-The key for each dictionary entry is the identification (id), as defined by the SBML specification, for the component. By printing one of these dictionaries keys, you can see the id of each compartment, parameter, or species.
-
-```python
-# get the dictionary keys for the IDs of the species in the model
-print(modelInstance.s.keys())
-# get the dictionary keys for the IDs of the compartments in the model
-print(modelInstance.c.keys())
[…]
