### 2019-09-02 46dec41
Added parent directory to path
