### 2019-09-02 8a8dae2
Couldn't get documentation to generate from readme and tutorial in parent directory. Recopied tutorial and readme files back
