+++ source/conf.py
@@ -20,7 +20,7 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../sbmltoodepy'))
-sys.path.insert(0, os.path.abspath('..'))
+
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
@@ -63,9 +63,9 @@ author = 'Steve M Ruggiero, Ashlee Nicole Ford Versypt'
 # built documents.
 #
 # The short X.Y version.
-version = '0.0'
+version = '1.0'
 # The full version, including alpha/beta/rc tags.
-release = '0.0'
+release = '1.0.2'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
