### 2019-09-03 SMRuggiero/sbmltoodepy#4 @AnabelSMRuggiero
We have added in both the readme and tutorial to our documentation page. While we see the value in a troubleshooting section of the documentation, we think we need user feedback before creating it. We plan on adding a changelog section to the documentation as part of finishing addressing reviewer feedback.
