### 2019-09-03 b0fd027
Fixed logic errors involving reaction parameters when generating reaction class and refomrating function mathstrings to use python functions.
