### 2019-09-03 289c47b
Fixed modelclasses submodel not being imported within the module by default (the TestPackage function was importing the submodule).
