+++ sbmltoodepy/__init__.py
@@ -1,2 +1,3 @@
-name = "sbmltopyode"
+name = "sbmltoodepy"
 from .utilities import ParseAndCreateModel
+import sbmltoodepy.modelclasses
