### 2019-09-08 #1643 @ashleefv
# Responses to Reviewer 1's comments (Issue 1):

### General Comments
We now specify that the supported versions of python are 3.5, 3.6, and 3.7.

Added comment to README.md about installing from source.

### Testing

PyTest could be a good fit for our project, but we expect to include if we have a major update in the future.

### Software checked on:
We now specify that the supported versions of python are 3.5, 3.6, and 3.7.

### Software style
Removed all of the "import *" statements. While the updated code on this github repo has been updated, we plan on updating the package on PyPI once all reviewer comments have been addressed.

### Running the provided Tutorial.md
We addressed the comments by

- clarifying the case when cloning the repo and pip install make sense together (use on google colab)
- fixing pip install command
- hosting a copy of SBMLtoODEpy.ipynb in the repo
###  Community guidelines

We do not see the scope of the project for others' to make pull requests or suggest features.
### Paper Clarifications
In our paper, we compare our work to two other pieces of software.

COPASI- The comparison was intended simply to validate the results of the models generated by SBMLtoODEpy. COPASI uses ODEPACK for its Time Course calculations, which is the same library that scipy's odeint function uses. If the calculations agree, then that suggests SBMLtoODEpy has generated the correct equations. Furthermore, the targeted niche for SBMLtoODEpy is not the same as COPASI.

Systems Biology Format Converter (SBFC)- SBFC is a better software to compare SBMLtoODEpy to. SBFC converts SBML models into various programming and modeling languages, but notably excludes Python. This gap SBFC's functionality is specifically the target niche of SBMLtoODEpy. Additionally, SBMLtoODEpy is designed to generate code that can be incorporated into other python projects. The Octave/MATLAB scripts generated by SBFC implement the model that neither accept arguments nor return any values. SBMLtoODEpy generates code that implements the model as a class with most methods and members intended to be accessible to other code.

The paper has been updated to better convey the scope and intent of the comparisons.
