### 2019-09-08 SMRuggiero/sbmltoodepy#2 @ashleefv
Since we expect users to take and modify or use the code generated by SBMLtoODEpy in ways we can't predict, the burden of applying SED-ML or KiSAO will have to be on the user. Additionally most SBML components are stored as entries in dictionaries that are members of the class implementing the model. Currently rules and initial assignments are not implemented this way and annotations stored in the SBML model are not retained. We plan to address these in a future update.

We updated the paper to expand on the comparison to COPASI. My comment to reviewer #1 addresses the comparison to COPASI and Systems Biology Format Converter in more depth.
