### 2019-09-08 SMRuggiero/sbmltoodepy#3 @ashleefv
We are looked into the suggestion and are currently not interested in using Travis-CI. Travis-CI does not play well with PyPI accounts with non-alphanumeric characters in its password. Additionally, any updates to SBMLtoODEpy will change the code generated by it and may require different interfacing with user code. Because of this, we want only want a new version of our code pushed when we are absolutely certain its ready for public use. Building and pushing a new version of our code to PyPI is already a simple process with setuptools and twine.

We did set up a code ocean capsule and tested our code on Ubuntu 16.04 and 18.04. We're not planning on publishing the capsule as we do not believe it to be a good fit to publish this project on code ocean. However, we are glad the reviewer provided this suggestion because it may be a good fit for some of our other projects that involve simulation experiments.
