### 2019-09-08 #1643 @ashleefv
# Responses to Reviewer 2's comments
## Comment 1 (Issue 2)
Since we expect users to take and modify or use the code generated by SBMLtoODEpy in ways we can't predict, the burden of applying SED-ML or KiSAO will have to be on the user. Additionally most SBML components are stored as entries in dictionaries that are members of the class implementing the model. Currently rules and initial assignments are not implemented this way and annotations stored in the SBML model are not retained. We plan to address these in a future update.

We updated the paper to expand on the comparison to COPASI. My comment to reviewer #1 addresses the comparison to COPASI and Systems Biology Format Converter in more depth.

## Comment 2 (Issue 3)

We looked into the suggestion and are currently not interested in using Travis-CI. Travis-CI does not play well with PyPI accounts with non-alphanumeric characters in its password. Additionally, any updates to SBMLtoODEpy will change the code generated by it and may require different interfacing with user code. Because of this, we want only want a new version of our code pushed when we are absolutely certain its ready for public use. Building and pushing a new version of our code to PyPI is already a simple process with setuptools and twine.

We did set up a code ocean capsule and tested our code on Ubuntu 16.04 and 18.04. We're not planning on publishing the capsule as we do not believe it to be a good fit to publish this project on code ocean. However, we are glad the reviewer provided this suggestion because it may be a good fit for some of our other projects that involve simulation experiments.

## Comment 3 (Issue 4)
We have added in both the readme and tutorial to our documentation page. While we see the value in a troubleshooting section of the documentation, we think we need user feedback before creating it. We plan on adding a changelog section to the documentation as part of finishing addressing reviewer feedback, if future edits are made.
