+++ paper.bib
@@ -176,11 +176,13 @@ doi={10.1093/bioinformatics/btl485},
 url={https://doi.org/10.1093/bioinformatics/btl485}
 }
 
-@article{hindmarsh1983,
-  title={ODEPACK, a systematized collection of ODE solvers},
+@incollection{hindmarsh1983,
+  title={{ODEPACK}, a systematized collection of {ODE} solvers},
   author={Hindmarsh, Alan C},
-  journal={Scientific computing},
+  booktitle={{Scientific Computing: Applications of Mathematics and Computing to the Physical Sciences}},
   pages={55--64},
+  editor = {Stepleman, R. S. and Carver, M. and Peskin, R. and Ames, W. F. and Vichnevetsky, R.},
   year={1983},
-  publisher={North-Holland}
+  publisher={North-Holland},
+  address = {New York}
 }
