### 2019-09-08 #1643 @mgymrek
Thanks @ashleefv for addressing these comments.
 @SirSharpest, @marouenbg, could you review these edits and update the checklists above accordingly?

Also, I noticed you mentioned you mentioned for community guidelines that you do not envision contributions from others to the project. However this is an important aspect of open source projects. Even if you do not envision new features, it's possible a user could find a bug and want to contribute a fix for example. To address this point you could simply add a blurb to the readme suggesting users to contact you or submit a pull request if they would like to contribute a change.
