### 2019-09-09 #1643 @marouenbg
I just checked all the boxes on my side. The authors did address the comments either through new commits and/or provided reasonable explanations for sticking to their initial choices. Great job!
