### 2019-09-12 #1643 @mgymrek
@ashleefv We are almost there, just a couple more steps. I have gone over the proof and have the following minor comments:

* "models that not easy" -> "models that are not easy"
* The Hindmarsh et al. reference I don't see a DOI in the proof

After fixing these, please make a Zenodo archive, ensuring the title + author list match that in the paper and that the version number is correct.
