+++ paper.md
@@ -29,10 +29,10 @@ SBML defines models independent of the software used to create the model, which 
 Programs for converting models into code exist for a variety of software languages but have key limitations.
 For example, the Systems Biology Format Converter (SBFC) is a suite of tools used to convert SBML models into scripts in multiple modeling and programming languages [@rodriguez2016].
 The BioModels Database uses SBFC to automate conversion of a large library of SBML models into other formats that can be downloaded by users [@chelliah2015; @glont2018].
-SBFC does not include a tool for generating Python scripts. For the languages SBFC supports, the codes generated are stand alone implementations of models that not easy to integrate into other projects.  
+SBFC does not include a tool for generating Python scripts. For the languages SBFC supports, the codes generated are stand alone implementations of models that are not easy to integrate into other projects.  
 
 The present work describes a software program called SBMLtoODEpy that we developed to address these limitations by enabling conversion of SBML models into Python classes that can be rapidly incorporated into biomedical systems modeling projects written in Python,
-such as the multiscale simulation platform CompuCell3D, or used directly in Python. The Octave and MATLAB code generated by SBFC implement SBML models as a function that accepts no input and doesn't return any values.
+such as the multiscale simulation platform CompuCell3D, or used directly in Python. The Octave and MATLAB code generated by SBFC implement SBML models as a function that accepts no input and does not return any values.
 SBMLtoODEpy generates code that uses Python classes to create code that users write their own code to interface with.
 The program aims to accelerate construction of multiscale models that import and reuse published SBML models,
 many of which are available in the BioModels Database at https://www.ebi.ac.uk/biomodels/
