### 2019-09-17 #1643 @mgymrek
Hi @ashleefv sorry for the delayed response, yes that is fine. You can now proceed with making the Zenodo archive. Thanks!
