+++ paper.bib
@@ -184,5 +184,8 @@ url={https://doi.org/10.1093/bioinformatics/btl485}
   editor = {Stepleman, R. S. and Carver, M. and Peskin, R. and Ames, W. F. and Vichnevetsky, R.},
   year={1983},
   publisher={North-Holland},
-  address = {New York}
+  address = {New York},
+  url = {https://computing.llnl.gov/casc/nsde/pubs/u88007.pdf}  
 }
+
+
