### 2019-09-18 730ec4a
Incremented version number to 1.0.3
