### 2019-09-18 bd0b827
Updated the README.md file for documentation generation
