### 2019-09-18 6afb46a
Updated DOI in readme
