### 2019-09-19 #1643 @ashleefv
@mgymrek We have updated the Hindmarsh url, fixed the noted typo, finalized the revised pypi release, and created a new zenodo doi for this updated release: https://zenodo.org/record/3441677#.XYLMoyhKjZs
