### 2019-09-19 94fd68f
included link for ipynb
