### 2019-09-19 7b4a396
Post release to fix DOI on PyPI
