### 2020-09-03 #1643 @ashleefv
I recently got the feedback that our acronym in the first word of the title reversed the B & M. It should be SBMLtoODEpy. Additionally this accidental reversal of the letters appeared one other time int he document. Can this be corrected?
