### 2020-09-03 #1643 @ashleefv
@mgymrek We aren't sure if our title issue and internal typo can be updated. We accidentally switched the letters B&M in the title of our paper and in one other place (I'm fine if that internal typo is not corrected, but the title would be best to fix.). I never caught the error because it blended together for me visually. Someone just commented on our github repository pointing out the issue.
