### 2020-09-04 #1643 @ashleefv
@arfon I have updated paper.md in the repository: https://github.com/SMRuggiero/sbmltoodepy
