### 2020-09-06 #1643 @arfon
@ashleefv - this was fixed in https://github.com/openjournals/joss-papers/commit/5f08ae368b4f81d8c7783bd4be821829cb0a3335 although the paper might take a few hours to show as updated on the JOSS website as there's caching in place.
